integ-tests: improve ad integration tests to test LDAPS with certificate verification
LDAPS connection has been tested, but without certificate verification. This commit create a self-signed certificate and instruct sssd agent to trust the certificate.

The reason for changing the `store_secret_in_secret_manager` from troposphere CloudFormation to boto3 calls is that CloudFormation only stores SecretString, but not SecretBytes

This commit also reduce the instance size used for the test.

+++ tests/integration-tests/tests/ad_integration/test_ad_integration.py
@@ -17,6 +17,7 @@ import os as os_lib
 import random
 import re
 import string
+import time
 import zipfile
 from collections import defaultdict
 
@@ -30,8 +31,6 @@ from OpenSSL.crypto import FILETYPE_PEM, TYPE_RSA, X509, dump_certificate, dump_
 from remote_command_executor import RemoteCommandExecutor
 from retrying import retry
 from time_utils import seconds
-from troposphere import Template
-from troposphere.secretsmanager import Secret
 from utils import generate_stack_name
 
 from tests.ad_integration.cluster_user import ClusterUser
@@ -51,20 +50,27 @@ def get_infra_stack_outputs(stack_name):
     }
 
 
-def get_ad_config_param_vals(directory_stack_name, nlb_stack_name, password_secret_arn):
+def get_infra_stack_parameters(stack_name):
+    cfn = boto3.client("cloudformation")
+    return {
+        entry.get("ParameterKey"): entry.get("ParameterValue")
+        for entry in cfn.describe_stacks(StackName=stack_name)["Stacks"][0]["Parameters"]
+    }
+
+
+def get_ad_config_param_vals(directory_stack_outputs, nlb_stack_parameters, password_secret_arn, ldap_tls_ca_cert):
     """Return a dict used to set values for config file parameters."""
-    directory_stack_outputs = get_infra_stack_outputs(directory_stack_name)
-    nlb_stack_outputs = get_infra_stack_outputs(nlb_stack_name)
     ldap_search_base = ",".join(
         [f"dc={domain_component}" for domain_component in directory_stack_outputs.get("DomainName").split(".")]
     )
     read_only_username = directory_stack_outputs.get("ReadOnlyUserName")
     return {
-        "ldaps_uri": nlb_stack_outputs.get("LDAPSURL"),
+        "ldaps_uri": nlb_stack_parameters.get("DomainName"),
         "ldap_search_base": ldap_search_base,
         # TODO: is the CN=Users portion of this a valid assumption?
         "ldap_default_bind_dn": f"CN={read_only_username},CN=Users,{ldap_search_base}",
         "password_secret_arn": password_secret_arn,
+        "ldap_tls_ca_cert": ldap_tls_ca_cert,
     }
 
 
@@ -114,31 +120,36 @@ def zip_dir(path):
     return file_out
 
 
-@pytest.fixture(scope="class")
-def store_secret_in_secret_manager(request, region, cfn_stacks_factory):
+@pytest.fixture(scope="module")
+def store_secret_in_secret_manager(request, cfn_stacks_factory):
 
-    secret_stack_name = generate_stack_name("integ-tests-secret", request.config.getoption("stackname_suffix"))
+    secret_arns = {}
 
-    def _store_secret(secret):
-        template = Template()
-        template.set_version("2010-09-09")
-        template.set_description("stack to store a secret string")
-        template.add_resource(Secret("Secret", SecretString=secret))
-        stack = CfnStack(
-            name=secret_stack_name,
-            region=region,
-            template=template.to_json(),
-        )
-        cfn_stacks_factory.create_stack(stack)
-        return stack.cfn_resources["Secret"]
+    def _store_secret(region, secret_string=None, secret_binary=None):
+        secrets_manager_client = boto3.client("secretsmanager")
+        if secret_string is None and secret_binary is None:
+            logging.error("secret string and scecret binary can not both be empty")
+        secret_name = generate_stack_name("integ-tests-secret", request.config.getoption("stackname_suffix"))
+        if secret_string:
+            secret_arn = secrets_manager_client.create_secret(Name=secret_name, SecretString=secret_string)["ARN"]
+        else:
+            secret_arn = secrets_manager_client.create_secret(Name=secret_name, SecretBinary=secret_binary)["ARN"]
+        if secret_arns.get(region):
+            secret_arns[region].append(secret_arn)
+        else:
+            secret_arns[region] = [secret_arn]
+        return secret_arn
 
     yield _store_secret
 
     if request.config.getoption("no_delete"):
-        logging.info("Not deleting stack %s because --no-delete option was specified", secret_stack_name)
+        logging.info("Not deleting stack secrets because --no-delete option was specified")
     else:
-        logging.info("Deleting stack %s", secret_stack_name)
-        cfn_stacks_factory.delete_stack(secret_stack_name, region)
+        for region, secrets in secret_arns.items():
+            secrets_manager_client = boto3.client("secretsmanager", region_name=region)
+            for secret_arn in secrets:
+                logging.info("Deleting secrete %s", secret_arn)
+                secrets_manager_client.delete_secret(SecretId=secret_arn)
 
 
 def _create_directory_stack(cfn_stacks_factory, request, directory_type, test_resources_dir, region, vpc_stack):
@@ -218,7 +229,16 @@ def _populate_directory_with_users(directory_stack, num_users_to_create, region)
     logging.info("Creation of %s users in directory service completed", str(num_users_to_create))
 
 
-def _create_nlb_stack(cfn_stacks_factory, request, directory_stack, region, test_resources_dir, certificate_arn):
+def _create_nlb_stack(
+    cfn_stacks_factory,
+    request,
+    directory_stack,
+    region,
+    test_resources_dir,
+    certificate_arn,
+    certificate_secret_arn,
+    domain_name,
+):
     nlb_stack_template_path = os_lib.path.join(test_resources_dir, "NLB_SimpleAD.yaml")
     nlb_stack_name = generate_stack_name(
         "integ-tests-MultiUserInfraStackNLB", request.config.getoption("stackname_suffix")
@@ -255,6 +275,14 @@ def _create_nlb_stack(cfn_stacks_factory, request, directory_stack, region, test
                     "ParameterKey": "SimpleADSecIP",
                     "ParameterValue": directory_stack.cfn_outputs["DirectoryDnsIpAddresses"].split(",")[1],
                 },
+                {
+                    "ParameterKey": "CertificateSecretArn",
+                    "ParameterValue": certificate_secret_arn,
+                },
+                {
+                    "ParameterKey": "DomainName",
+                    "ParameterValue": domain_name,
+                },
             ],
         )
     cfn_stacks_factory.create_stack(nlb_stack)
@@ -262,11 +290,10 @@ def _create_nlb_stack(cfn_stacks_factory, request, directory_stack, region, test
     return nlb_stack
 
 
-def _generate_certificate():
+def _generate_certificate(common_name):
     key = crypto.PKey()
     key.generate_key(TYPE_RSA, 2048)
     crt = X509()
-    common_name = "ldap.simplead.multiuser.pcluster"
     crt.get_subject().commonName = common_name
     crt.get_issuer().commonName = common_name
     now = datetime.datetime.now()
@@ -277,20 +304,26 @@ def _generate_certificate():
     crt.set_serial_number(random.randrange(1, 99999))
     crt.set_pubkey(key)
     crt.sign(key, "sha256")
-    certificate_arn = boto3.client("acm").import_certificate(
-        Certificate=dump_certificate(FILETYPE_PEM, crt), PrivateKey=dump_privatekey(FILETYPE_PEM, key)
-    )["CertificateArn"]
-    return certificate_arn
+    certificate = dump_certificate(FILETYPE_PEM, crt)
+    private_key = dump_privatekey(FILETYPE_PEM, key)
+    certificate_arn = boto3.client("acm").import_certificate(Certificate=certificate, PrivateKey=private_key)[
+        "CertificateArn"
+    ]
+    return certificate_arn, certificate
 
 
 @pytest.fixture(scope="module")
-def directory_factory(request, cfn_stacks_factory, vpc_stacks):
+def directory_factory(request, cfn_stacks_factory, vpc_stacks, store_secret_in_secret_manager):  # noqa: C901
     # TODO: use external data file and file locking in order to share directories across processes
     created_directory_stacks = defaultdict(dict)
     created_certificates = defaultdict(dict)
 
     def _directory_factory(
-        existing_directory_stack_name, existing_nlb_stack_name, directory_type, test_resources_dir, region
+        existing_directory_stack_name,
+        existing_nlb_stack_name,
+        directory_type,
+        test_resources_dir,
+        region,
     ):
         if existing_directory_stack_name:
             directory_stack_name = existing_directory_stack_name
@@ -315,10 +348,19 @@ def directory_factory(request, cfn_stacks_factory, vpc_stacks):
             nlb_stack_name = created_directory_stacks.get(region, {}).get("nlb")
             logging.info("Using NLB stack named %s created by another test", nlb_stack_name)
         else:
-            certificate_arn = _generate_certificate()
+            common_name = f"{directory_type.lower()}.multiuser.pcluster"
+            certificate_arn, certificate = _generate_certificate(common_name)
+            certificate_secret_arn = store_secret_in_secret_manager(region, secret_binary=certificate)
             created_certificates[region] = certificate_arn
             nlb_stack_name = _create_nlb_stack(
-                cfn_stacks_factory, request, directory_stack, region, test_resources_dir, certificate_arn
+                cfn_stacks_factory,
+                request,
+                directory_stack,
+                region,
+                test_resources_dir,
+                certificate_arn,
+                certificate_secret_arn,
+                common_name,
             ).name
             created_directory_stacks[region]["nlb"] = nlb_stack_name
         return directory_stack_name, nlb_stack_name
@@ -340,7 +382,8 @@ def directory_factory(request, cfn_stacks_factory, vpc_stacks):
                 cfn_stacks_factory.delete_stack(stack_name, region)
 
     for region, certificate_arn in created_certificates.items():
-        boto3.client("acm", region_name=region).delete_certificate(CertificateArn=certificate_arn)
+        if request.config.getoption("no_delete"):
+            boto3.client("acm", region_name=region).delete_certificate(CertificateArn=certificate_arn)
 
 
 def _run_user_workloads(users, test_datadir, remote_command_executor):
@@ -501,13 +544,32 @@ def test_ad_integration(
     )
     directory_stack_outputs = get_infra_stack_outputs(directory_stack_name)
     ad_user_password = directory_stack_outputs.get("UserPassword")
-    password_secret_arn = store_secret_in_secret_manager(directory_stack_outputs.get("AdminPassword"))
-    config_params.update(get_ad_config_param_vals(directory_stack_name, nlb_stack_name, password_secret_arn))
+    password_secret_arn = store_secret_in_secret_manager(
+        region, secret_string=directory_stack_outputs.get("AdminPassword")
+    )
+    nlb_stack_parameters = get_infra_stack_parameters(nlb_stack_name)
+    ldap_tls_ca_cert = "/opt/parallelcluster/shared/directory_service/certificate.crt"
+    config_params.update(
+        get_ad_config_param_vals(directory_stack_outputs, nlb_stack_parameters, password_secret_arn, ldap_tls_ca_cert)
+    )
     cluster_config = pcluster_config_reader(**config_params)
     cluster = clusters_factory(cluster_config)
 
+    certificate_secret_arn = nlb_stack_parameters.get("CertificateSecretArn")
+    certificate = boto3.client("secretsmanager").get_secret_value(SecretId=certificate_secret_arn)["SecretBinary"]
+    with open(test_datadir / "certificate.crt", "wb") as f:
+        f.write(certificate)
+
     # Publish compute node count metric every minute via cron job
     # TODO: use metrics reporter from the benchmarks module
+    remote_command_executor = RemoteCommandExecutor(cluster)
+    remote_command_executor.run_remote_command(
+        f"sudo cp certificate.crt {ldap_tls_ca_cert} && sudo service sssd restart",
+        additional_files=[test_datadir / "certificate.crt"],
+    )
+    time.sleep(600)
+    # TODO: we have to sleep for 10 minutes to wait for the SSSD agent use the newly placed certificate.
+    #  We should look for other methods to let the SSSD agent use the new certificate more quickly
     remote_command_executor = RemoteCommandExecutor(cluster)
     metric_publisher_script = "publish_compute_node_count_metric.sh"
     remote_metric_publisher_script_path = f"/shared/{metric_publisher_script}"
